feat: add Tour rating essentials

+++ src/Modules/Tours/Explorer.Tours.Core/Domain/RepositoryInterfaces/ITourExecutionRepository.cs
@@ -9,7 +9,7 @@ namespace Explorer.Tours.Core.Domain.RepositoryInterfaces
 {
     public interface ITourExecutionRepository
     {
-        public TourExecution Get(int id);
+        public TourExecution Get(long id);
         public TourExecution Create(TourExecution execution);
         public TourExecution Update(TourExecution execution);
         public void Delete(int id);
